Split database interface to Database and Config.
Database to save marks,
Config to read questions and configuration.

+++ learned-cat-interfaces/src/settings.rs
@@ -20,7 +20,7 @@ pub struct TestSettings {
     pub test_duration_minutes: i64,
 
     #[serde(default)]
-    pub number_of_attempts: usize,
+    pub number_of_attempts: u32,
 
     /// Castumization
     #[serde(default)]
